requirements.txt
fastapi
uvicorn
streamlit
pydantic
python-dotenv
openai
requests